ocr_function/requirements.txt
azure-functions
azure-storage-blob
azure-ai-formrecognizer
sendgrid